+++ src/components/conversation.tsx
@@ -34,7 +34,7 @@ import VisibilitySensor from "react-visibility-sensor";
 import { useHistory, useIsSmallScreen, useUser } from "./AppContextProvider";
 import PageBottom from "./conversation/pageBottom";
 import PageSelect from "./conversation/pageSelect";
-import useBtns from "./conversation/functions/btns";
+import useBtns from "../hooks/conversation/btns";
 import { PhotoProvider } from "react-photo-view";
 import {
     useCBottom,
@@ -51,11 +51,11 @@ import {
     useUpdating,
     useVotes,
 } from "./conversation/ConversationContext";
-import { useUpdate } from "./conversation/functions/update";
-import useFirstFetch from "./conversation/functions/firstfetch";
-import useChangePage from "./conversation/functions/changePage";
-import useOnScroll from "./conversation/functions/onScroll";
-import useOnVisibilityChange from "./conversation/functions/onVisibilityChange";
+import { useUpdate } from "../hooks/conversation/update";
+import useFirstFetch from "../hooks/conversation/firstfetch";
+import useChangePage from "../hooks/conversation/changePage";
+import useOnScroll from "../hooks/conversation/onScroll";
+import useOnVisibilityChange from "../hooks/conversation/onVisibilityChange";
 import FloatingEditor from "./floatingEditor";
 import Gallery from "./conversation/gallery";
 import Dock from "./dock";